The Detail and Create components have been refactored.
